+++ doc/Changelog.md
@@ -2,6 +2,9 @@
 
 # git master
 
+* [7](https://github.com/BlueBrain/Keyv/pull/7):
+  Fix takeValues on OS X or latest memcached
+
 # Release 1.0 (09-12-2016)
 
 * [2](https://github.com/BlueBrain/Keyv/pull/2):