+++ doc/Changelog.md
@@ -2,6 +2,8 @@
 
 # git master
 
+* [8](https://github.com/BlueBrain/Keyv/pull/8):
+  Add compression for memcached
 * [7](https://github.com/BlueBrain/Keyv/pull/7):
   Fix takeValues on OS X or latest memcached